Creating a trigger is now possible
