+++ HomeApp/app/src/main/java/com/HomeApp/screens/RoutinesScreen.kt
@@ -60,7 +60,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.HomeApp.ui.composables.RoutinesTitleBar
-import com.HomeApp.ui.composables.RoutinesTitleBarItem
+import com.HomeApp.ui.composables.TopTitleBarItem
 import com.HomeApp.ui.theme.DarkRed
 import com.HomeApp.ui.theme.FadedLightGrey
 import com.HomeApp.ui.theme.GhostWhite
@@ -86,7 +86,7 @@ fun RoutinesScreen(
 
     Scaffold(
         topBar = {
-            RoutinesTitleBar(item = RoutinesTitleBarItem.Routines, navController = navController)
+            RoutinesTitleBar(item = TopTitleBarItem.Routines, navController = navController)
         },
         content = {
             LazyColumn(
