+++ HomeApp/app/src/main/java/com/HomeApp/ui/composables/DeviceCard.kt
@@ -1,17 +1,43 @@
 package com.HomeApp.ui.composables
 
 import android.content.Context
-import android.util.Log
 import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.*
+import androidx.compose.material.AlertDialog
+import androidx.compose.material.Button
+import androidx.compose.material.ButtonDefaults
+import androidx.compose.material.Icon
+import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.BrokenImage
 import androidx.compose.material.icons.filled.DoorFront
 import androidx.compose.material.icons.filled.Lightbulb
-import androidx.compose.material.icons.outlined.*
-import androidx.compose.runtime.*
+import androidx.compose.material.icons.outlined.CompareArrows
+import androidx.compose.material.icons.outlined.Lock
+import androidx.compose.material.icons.outlined.LockOpen
+import androidx.compose.material.icons.outlined.MusicNote
+import androidx.compose.material.icons.outlined.RestartAlt
+import androidx.compose.material.icons.outlined.Sensors
+import androidx.compose.material.icons.outlined.SmartScreen
+import androidx.compose.material.icons.outlined.SurroundSound
+import androidx.compose.material.icons.outlined.TextIncrease
+import androidx.compose.material.icons.outlined.Window
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
@@ -21,7 +47,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
-import com.HomeApp.ui.navigation.Sensor
+import com.HomeApp.screens.SelectedItems
+import com.HomeApp.ui.navigation.ChooseItems
 import com.HomeApp.ui.theme.RaminGrey
 import com.HomeApp.util.ApiConnector
 import com.HomeApp.util.ApiResult
@@ -83,7 +110,11 @@ fun DeviceCard(
         Button(
             onClick = {
                 when (deviceItem.get("type") as String) {
-                    "sensor" -> navController.navigate(Sensor.route)
+                    "sensor" -> {
+                        SelectedItems.setSensorId(deviceItem.id)
+                        SelectedItems.setType("sensor")
+                        navController.navigate(ChooseItems.route)
+                    }
                     in stateList -> {
                         changeState(
                             context = context,
@@ -123,7 +154,9 @@ fun DeviceCard(
         Button(
             onClick = {
                 if (deviceItem.get("type") as String == "sensor") {
-                    navController.navigate(Sensor.route)
+                    SelectedItems.setSensorId(deviceItem.id)
+                    SelectedItems.setType("sensor")
+                    navController.navigate(ChooseItems.route)
                 } else {
                     changeState(
                         context = context,
