+++ HomeApp/app/src/main/java/com/HomeApp/ui/composables/TopTitleBar.kt
@@ -27,15 +27,16 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.HomeApp.screens.SelectedItems
+import com.HomeApp.ui.navigation.ChooseActions
 import com.HomeApp.ui.navigation.Devices
 import com.HomeApp.ui.navigation.Home
-import com.HomeApp.ui.navigation.Sensor
 
 @Composable
 fun RoutinesTitleBar(
-    item: RoutinesTitleBarItem,
+    item: TopTitleBarItem,
     navController: NavController
 ) {
+    val type = SelectedItems.getType()
     val showDialog = remember { mutableStateOf(false) }
     if (showDialog.value) {
         Dialog(
@@ -53,7 +54,21 @@ fun RoutinesTitleBar(
         ) {
             IconButton(
                 modifier = Modifier.weight(1f),
-                onClick = { navController.navigate(item.routeLeftButton) }
+                onClick = {
+                    if (item == TopTitleBarItem.ChooseItems) {
+                        if (type == "sensor") {
+                            navController.navigate(Devices.route)
+                            return@IconButton
+                        }
+                    }
+                    if (item == TopTitleBarItem.Finish) {
+                        if (type == "sensor") {
+                            navController.navigate(ChooseActions.route)
+                            return@IconButton
+                        }
+                    }
+                    navController.navigate(item.routeLeftButton)
+                }
             ) {
                 Icon(
                     modifier = Modifier.scale(2.2f),
@@ -62,8 +77,8 @@ fun RoutinesTitleBar(
                 )
             }
             Text(
-                text = if (item == RoutinesTitleBarItem.ChooseItems) {
-                    if (SelectedItems.getType()[0]) "Devices" else "Groups"
+                text = if (item == TopTitleBarItem.ChooseItems) {
+                    if (type == "group") "Groups" else "Devices"
                 } else {
                     item.title
                 },
@@ -122,14 +137,14 @@ private fun Dialog(
     )
 }
 
-sealed class RoutinesTitleBarItem(
+sealed class TopTitleBarItem(
     val title: String,
     val iconLeft: ImageVector,
     val routeLeftButton: String,
     val iconRight: ImageVector,
     val routeRightButton: String
 ) {
-    object Routines : RoutinesTitleBarItem(
+    object Routines : TopTitleBarItem(
         title = "Routines",
         iconLeft = Icons.Rounded.ArrowBack,
         routeLeftButton = Home.route,
@@ -137,7 +152,7 @@ sealed class RoutinesTitleBarItem(
         routeRightButton = com.HomeApp.ui.navigation.ChooseType.route
     )
 
-    object ChooseType : RoutinesTitleBarItem(
+    object ChooseType : TopTitleBarItem(
         title = "Choose",
         iconLeft = Icons.Rounded.ArrowBack,
         routeLeftButton = com.HomeApp.ui.navigation.Routines.route,
@@ -145,7 +160,7 @@ sealed class RoutinesTitleBarItem(
         routeRightButton = com.HomeApp.ui.navigation.Routines.route
     )
 
-    object ChooseItems : RoutinesTitleBarItem(
+    object ChooseItems : TopTitleBarItem(
         title = "Items",
         iconLeft = Icons.Rounded.ArrowBack,
         routeLeftButton = com.HomeApp.ui.navigation.ChooseType.route,
@@ -153,7 +168,7 @@ sealed class RoutinesTitleBarItem(
         routeRightButton = com.HomeApp.ui.navigation.Routines.route
     )
 
-    object ChooseActions : RoutinesTitleBarItem(
+    object ChooseActions : TopTitleBarItem(
         title = "Actions",
         iconLeft = Icons.Rounded.ArrowBack,
         routeLeftButton = com.HomeApp.ui.navigation.ChooseItems.route,
@@ -161,7 +176,7 @@ sealed class RoutinesTitleBarItem(
         routeRightButton = com.HomeApp.ui.navigation.Routines.route
     )
 
-    object ChooseSchedule : RoutinesTitleBarItem(
+    object ChooseSchedule : TopTitleBarItem(
         title = "Schedule",
         iconLeft = Icons.Rounded.ArrowBack,
         routeLeftButton = com.HomeApp.ui.navigation.ChooseActions.route,
@@ -169,35 +184,11 @@ sealed class RoutinesTitleBarItem(
         routeRightButton = com.HomeApp.ui.navigation.Routines.route
     )
 
-    object FinishRoutine : RoutinesTitleBarItem(
+    object Finish : TopTitleBarItem(
         title = "Summary",
         iconLeft = Icons.Rounded.ArrowBack,
         routeLeftButton = com.HomeApp.ui.navigation.ChooseSchedule.route,
         iconRight = Icons.Rounded.Close,
         routeRightButton = com.HomeApp.ui.navigation.Routines.route
-    )
-
-    object Sensor : RoutinesTitleBarItem(
-        title = "Sensor",
-        iconLeft = Icons.Rounded.ArrowBack,
-        routeLeftButton = Devices.route,
-        iconRight = Icons.Rounded.Close,
-        routeRightButton = Devices.route
-    )
-
-    object SensorActions : RoutinesTitleBarItem(
-        title = "Actions",
-        iconLeft = Icons.Rounded.ArrowBack,
-        routeLeftButton = com.HomeApp.ui.navigation.Sensor.route,
-        iconRight = Icons.Rounded.Close,
-        routeRightButton = Devices.route
-    )
-
-    object SensorFinish : RoutinesTitleBarItem(
-        title = "Finish",
-        iconLeft = Icons.Rounded.ArrowBack,
-        routeLeftButton = com.HomeApp.ui.navigation.SensorAction.route,
-        iconRight = Icons.Rounded.Close,
-        routeRightButton = Devices.route
     )
 }
