+++ HomeApp/app/src/main/java/com/HomeApp/ui/navigation/NavHost.kt
@@ -279,59 +279,17 @@ fun AnimatedAppNavHost(
             )
         }
 
-        // FINISH ROUTINE
+        // FINISH ROUTINE/SENSOR
         composable(
-            route = FinishRoutine.route,
+            route = Finish.route,
             enterTransition = { fadeIn(tween(defaultTween)) },
             popEnterTransition = { fadeIn(tween(defaultTween)) },
             exitTransition = { fadeOut(tween(defaultTween)) },
             popExitTransition = { fadeOut(tween(defaultTween)) }
         ) {
-            RoutineFinishScreen(
+            FinishScreen(
                 navController = navController,
-                OnSelfClick = { navController.navigateSingleTopTo(FinishRoutine.route) }
-            )
-        }
-
-        // SENSOR
-        composable(
-            route = Sensor.route,
-            enterTransition = { fadeIn(tween(defaultTween)) },
-            popEnterTransition = { fadeIn(tween(defaultTween)) },
-            exitTransition = { fadeOut(tween(defaultTween)) },
-            popExitTransition = { fadeOut(tween(defaultTween)) }
-        ) {
-            SensorScreen(
-                navController = navController,
-                OnSelfClick = { navController.navigateSingleTopTo(Sensor.route) }
-            )
-        }
-
-        // SENSOR ACTION
-        composable(
-            route = SensorAction.route,
-            enterTransition = { fadeIn(tween(defaultTween)) },
-            popEnterTransition = { fadeIn(tween(defaultTween)) },
-            exitTransition = { fadeOut(tween(defaultTween)) },
-            popExitTransition = { fadeOut(tween(defaultTween)) }
-        ) {
-            SensorActionScreen(
-                navController = navController,
-                OnSelfClick = { navController.navigateSingleTopTo(SensorAction.route) }
-            )
-        }
-
-        // FINISH SENSOR
-        composable(
-            route = SensorFinish.route,
-            enterTransition = { fadeIn(tween(defaultTween)) },
-            popEnterTransition = { fadeIn(tween(defaultTween)) },
-            exitTransition = { fadeOut(tween(defaultTween)) },
-            popExitTransition = { fadeOut(tween(defaultTween)) }
-        ) {
-            SensorFinishScreen(
-                navController = navController,
-                OnSelfClick = { navController.navigateSingleTopTo(SensorFinish.route) }
+                OnSelfClick = { navController.navigateSingleTopTo(Finish.route) }
             )
         }
     }
